+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Incompatible changes:
 * Other backwards incompatible function API changes and code refactoring.
 
 New features:
+* Add random_seed option to tranforming.permute_data()
 * Add bad_color parameter to heatmap() and derivative functions
 * Add more methods for MS1Experiment
 * Add q-values (correted p-values) to dsfdr and derivative functions. This is manifested in a new feature_metadata field ("qval") for results of diff_abundance() / correlation()